feat: refactor primitives, add delegation parsing fn
