+++ bolt-sidecar/Config.example.toml
@@ -1,0 +1,20 @@
+# ports
+port = 8000
+metrics_port = 3300
+
+# node urls
+execution_api_url = "http://localhost:8545"
+beacon_api_url = "http://localhost:5052"
+engine_api_url = "http://localhost:8551"
+
+# constraints options
+constraints_url = "http://localhost:3030"
+constraints_proxy_port = 18551
+
+# chain options
+chain = "kurtosis"
+slot_time = 2
+
+# signing options
+private_key = "0x359c156600e1f5715a58c9e09f17c8d04e7ee3a9f88b39f6e489ffca0148fabe"
+delegations_path = "./delegations.json"
